+++ src/extension.ts
@@ -6,6 +6,7 @@ import { FileSystemRouter } from './fileSystemRouter';
 import { Logging, setDebug } from './logging';
 import { Manager } from './manager';
 import { registerChatTools } from './chatTools';
+import { registerChatParticipant } from './chatParticipant';
 import { registerSearchProviders } from './searchProvider';
 import type { SSHPseudoTerminal } from './pseudoTerminal';
 import { ConfigTreeProvider, ConnectionTreeProvider } from './treeViewManager';
@@ -67,6 +68,7 @@ export function activate(context: vscode.ExtensionContext) {
   subscribe(vscode.workspace.registerFileSystemProvider('ssh', new FileSystemRouter(manager), { isCaseSensitive: true }));
   registerSearchProviders(manager.connectionManager, subscribe);
   registerChatTools(manager.connectionManager, subscribe);
+  registerChatParticipant(manager.connectionManager, context);
   subscribe(vscode.window.createTreeView('sshfs-configs', { treeDataProvider: new ConfigTreeProvider(), showCollapseAll: true }));
   const connectionTreeProvider = new ConnectionTreeProvider(manager.connectionManager);
   subscribe(vscode.window.createTreeView('sshfs-connections', { treeDataProvider: connectionTreeProvider, showCollapseAll: true }));